requirements.txt
flask
requests
python-dotenv
twilio
click
rich
tabulate
